Se agrego el Atomic Design para la feature/dashboard
Patrón de diseño Atómico terminado para la vista del tablero(DashBoardPage.tsx)
