+++ src/features/dashboard/presentation/ui/components/pages/DashBoardPage.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import { View } from "react-native";
+import DashBoardTemplate from "../templates/DashBoardTemplate";
+import RelationshipGraph from "../organisms/RelationshipGraph";
+
+export default function DashBoardPage () {
+    return (
+        <View style={{ flex: 1 }}>
+            <DashBoardTemplate>
+                <RelationshipGraph />
+            </DashBoardTemplate>
+        </View>
+    );
+}
