+++ src/features/dashboard/presentation/ui/components/templates/DashBoardTemplate.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+import { View } from 'react-native';
+import SafeLayout from '@/shared/components/layouts/SafeLayout';
+
+export default function DashBoardTemplate ({children}: {children: React.ReactNode}) {
+    return (
+        <SafeLayout>
+            <View
+                style={{
+                    flex:1,
+                    justifyContent: 'center',
+                    alignItems: 'center',
+                }}
+            >
+                {children}
+            </View>
+        </SafeLayout>
+    );
+}
