+++ src/shared/theme/Typography.ts
@@ -1,7 +1,4 @@
 export const typography = {
-    //peso de la fuente
-    fontWeightBold: 'bold' as const,
-
     //Tamaños para la fuente
     fontSizeXL: 26,
     fontSizeL:  20,
